refactor: Rename 'gktest' to 'gator'
Fixes #1722
