+++ pkg/gator/printer_go_test.go
@@ -1,4 +1,4 @@
-package gktest
+package gator
 
 import (
 	"errors"
