+++ pkg/gator/read_constraints.go
@@ -1,4 +1,4 @@
-package gktest
+package gator
 
 import (
 	"encoding/json"
