+++ pkg/gator/read_suites_test.go
@@ -1,4 +1,4 @@
-package gktest
+package gator
 
 import (
 	"errors"
