+++ pkg/gator/runner_integer_test.go
@@ -1,4 +1,4 @@
-package gktest
+package gator
 
 import (
 	"context"
