+++ pkg/gator/suite.go
@@ -1,4 +1,4 @@
-package gktest
+package gator
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
